fix error

+++ auth/djangoRest/expiredToken/main/middlewares.py
@@ -13,21 +13,22 @@ class MyMiddleware:
         
         if token is not None:
             dbToken = Token.objects.filter(key=token.split(" ")[1])
-            created = dbToken.first().created
-            # print("created: ", created)
+            if dbToken.count() != 0:
+                created = dbToken.first().created
+                # print("created: ", created)
 
-            utc_now = datetime.utcnow()
-            utc_now = utc_now.replace(tzinfo=pytz.utc)
-            # print("now: ", utc_now)
+                utc_now = datetime.utcnow()
+                utc_now = utc_now.replace(tzinfo=pytz.utc)
+                # print("now: ", utc_now)
 
-            delta = utc_now - created
-            print(delta)
+                delta = utc_now - created
+                print(delta)
 
-            if delta > timedelta(minutes=settings.EXPIRATION_IN_MIN):
-                print("expired!! signout!!")
-                dbToken.delete()
-            else:
-                print("OK, extend you")
-                dbToken.update(created = utc_now)
+                if delta > timedelta(minutes=settings.EXPIRATION_IN_MIN):
+                    print("expired!! signout!!")
+                    dbToken.delete()
+                else:
+                    print("OK, extend you")
+                    dbToken.update(created = utc_now)
 
         return self.get_response(request)
